amu_pay/ai/requirements.txt
# ============================================================================
# AI App Requirements for AMU Pay
# Django AI Chatbot with LangChain, Google Gemini, and Pinecone RAG
# ============================================================================
# 
# Installation: pip install -r requirements.txt
# If Windows file lock error: pip install --user -r requirements.txt
#
# Last Updated: October 31, 2025
# Python Version: 3.12+
# ============================================================================

# ============================================================================
# CORE LANGCHAIN FRAMEWORK
# ============================================================================
# Main LangChain packages for building AI agents and chains
# Version 0.2.x is stable and compatible with all integrations

langchain==0.2.7                    # Main LangChain framework
langchain-core==0.2.38              # Core abstractions (UPDATED for compatibility)
langchain-community==0.2.7          # Community integrations
langgraph==0.1.17                   # Graph-based agent workflows

# ============================================================================
# AI MODEL PROVIDERS
# ============================================================================
# Google Gemini AI (Primary LLM)

langchain-google-genai==1.0.10      # LangChain integration for Gemini
google-generativeai==0.7.2          # Google Generative AI SDK

# Alternative LLM Providers (Optional - Uncomment if needed)
# langchain-openai==0.2.8           # For OpenAI GPT models
# langchain-anthropic==0.2.8        # For Anthropic Claude models

# ============================================================================
# VECTOR DATABASE & RAG (Retrieval-Augmented Generation)
# ============================================================================
# Pinecone for semantic search and documentation retrieval

pinecone-client>=3.2.2,<4.0.0       # Pinecone vector database client (compatible with langchain-pinecone 0.1.1)
pinecone-plugin-assistant==1.8.0    # Assistant plugin for Pinecone
langchain-pinecone==0.1.1           # LangChain Pinecone integration (compatible with langchain-core 0.2.x)

# ============================================================================
# EMBEDDINGS & ML MODELS
# ============================================================================
# For converting text to vectors (semantic search)

langchain-huggingface==0.0.1        # LangChain HuggingFace integration
sentence-transformers==2.6.1        # Multilingual embeddings
                                     # Uses: paraphrase-multilingual-mpnet-base-v2

# ============================================================================
# DATA PROCESSING & VALIDATION
# ============================================================================

pydantic==2.10.3                    # Data validation and settings management
python-decouple==3.8                # Environment variable management (.env)
python-dotenv==1.0.1                # Needed by setup_pinecone.py (from dotenv import load_dotenv)
orjson==3.10.12                     # Fast JSON serialization

# ============================================================================
# HTTP & NETWORKING
# ============================================================================

requests==2.32.3                    # HTTP library (UPDATED for Pinecone compatibility)
httpx==0.27.2                       # Async HTTP client

# ============================================================================
# ASYNC & MESSAGING
# ============================================================================

asyncio-mqtt==0.16.2                # Async MQTT client

# ============================================================================
# LOGGING & MONITORING
# ============================================================================

structlog==24.4.0                   # Structured logging

# ============================================================================
# DEVELOPMENT & TESTING
# ============================================================================
# For testing the AI components

pytest==8.3.4                       # Testing framework
pytest-asyncio==0.24.0              # Async testing support
pytest-mock==3.14.0                 # Mocking for tests

# ============================================================================
# TYPE CHECKING (Optional)
# ============================================================================
# For static type checking with mypy

mypy==1.13.0                        # Static type checker
types-requests==2.32.0.20241016     # Type stubs for requests

# ============================================================================
# DOCUMENTATION (Optional - Uncomment if needed)
# ============================================================================
# For generating documentation

# sphinx==8.1.3                     # Documentation generator
# sphinx-rtd-theme==3.1.1           # Read the Docs theme

# ============================================================================
# ADDITIONAL LANGCHAIN TOOLS (Optional - Uncomment if needed)
# ============================================================================

# langchain-experimental==0.2.2     # Experimental chains and features

# ============================================================================
# DEPENDENCY NOTES
# ============================================================================
#
# Key Version Updates (from original):
# - langchain-core: 0.2.12 → 0.2.38 (for langgraph & langchain-google-genai)
# - langchain-google-genai: 1.0.3 → 1.0.10 (for langchain-core 0.2.x support)
# - google-generativeai: 0.5.4 → 0.7.2 (for langchain-google-genai 1.0.10)
# - requests: 2.31.0 → 2.32.3 (for pinecone-plugin-assistant)
# - python-dotenv → python-decouple (to match code imports)
#
# All versions tested and verified compatible on:
# - Python 3.12
# - Windows 10/11
# - October 31, 2025
#
# ============================================================================
# ENVIRONMENT VARIABLES REQUIRED
# ============================================================================
#
# Create a .env file in project root with:
#
# GEMINI_API_KEY=your_gemini_api_key_here
# PINECONE_API_KEY=your_pinecone_api_key_here
# PINECONE_INDEX_NAME=amu-pay-docs
#
# ============================================================================
